server/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==7.2.7
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
django-stubs==4.2.1
django-stubs-ext==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.14.1
executing==1.2.0
idna==3.4
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
mypy==1.3.0
mypy-extensions==1.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-extensions==4.6.2
urllib3==2.0.2
wcwidth==0.2.6
